chore: Make pure functions for A* algo
